# lib/api-client-react/src/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Proxy Portal API - Manages Vercel API keys and proxies AI model requests
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiKey {
  id: number;
  name: string;
  /** Last 4 characters of the key for identification */
  keyPreview: string;
  isEnabled: boolean;
  totalRequests: number;
  totalTokens: number;
  estimatedCostUsd: number;
  /** @nullable */
  lastUsedAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateKeyBody {
  /** Friendly name for the key */
  name: string;
  /** The Vercel API key (will be stored securely) */
  apiKey: string;
}

export interface UpdateKeyBody {
  name?: string;
  isEnabled?: boolean;
}

export interface UsageLog {
  id: number;
  keyId: number;
  keyName: string;
  model: string;
  promptTokens: number;
  completionTokens: number;
  totalTokens: number;
  estimatedCostUsd: number;
  statusCode: number;
  durationMs: number;
  createdAt: string;
}

export interface KeyStats {
  totalKeys: number;
  enabledKeys: number;
  totalRequests: number;
  totalTokens: number;
  totalCostUsd: number;
  recentActivity: UsageLog[];
}

export interface KeyUsageSummary {
  keyId: number;
  keyName: string;
  requests: number;
  tokens: number;
  costUsd: number;
  successRate: number;
}

export interface UsageSummary {
  byKey: KeyUsageSummary[];
  totalRequests: number;
  totalCostUsd: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListUsageParams = {
  /**
   * @nullable
   */
  keyId?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};