# lib/api-zod/src/generated/types/apiKey.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Proxy Portal API - Manages Vercel API keys and proxies AI model requests
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKey {
  id: number;
  name: string;
  /** Last 4 characters of the key for identification */
  keyPreview: string;
  isEnabled: boolean;
  totalRequests: number;
  totalTokens: number;
  estimatedCostUsd: number;
  /** @nullable */
  lastUsedAt: string | null;
  createdAt: string;
  updatedAt: string;
}