# lib/api-zod/src/generated/types/keyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Proxy Portal API - Manages Vercel API keys and proxies AI model requests
 * OpenAPI spec version: 0.1.0
 */

export interface KeyInput {
  /** Friendly name for the key */
  name: string;
  /** The Vercel API key (will be stored securely) */
  apiKey: string;
}