# lib/api-zod/src/generated/types/keyStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Proxy Portal API - Manages Vercel API keys and proxies AI model requests
 * OpenAPI spec version: 0.1.0
 */
import type { UsageLog } from './usageLog';

export interface KeyStats {
  totalKeys: number;
  enabledKeys: number;
  totalRequests: number;
  totalTokens: number;
  totalCostUsd: number;
  recentActivity: UsageLog[];
}