# lib/api-zod/src/generated/types/usageLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Proxy Portal API - Manages Vercel API keys and proxies AI model requests
 * OpenAPI spec version: 0.1.0
 */

export interface UsageLog {
  id: number;
  keyId: number;
  keyName: string;
  model: string;
  promptTokens: number;
  completionTokens: number;
  totalTokens: number;
  estimatedCostUsd: number;
  statusCode: number;
  durationMs: number;
  createdAt: string;
}